Adding rebar3 support for compiling .peg files
When using this dependency in with `rebar3`, I noticed the `.peg`
files were not compiling and therefore the `props` library was
missing a beam required to function. By adding the `rebar3`
neotoma plugin and also adding the `provider_hook` to use the plugin
on compile, the `.peg` file is correctly compiled and the `props`
library works again.